% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clar.R
\name{clar}
\alias{clar}
\title{Cumulative LGD Accuracy Ratio}
\usage{
clar(y, hx)
}
\arguments{
\item{y}{a vector of  realized values.}

\item{hx}{a vector of  predicted values.}
}
\value{
The function returns the CLAR for a vector of realized categories \code{y} and a vector of predicted categories \code{hx}.
}
\description{
Calculates for a vector of realized categories \code{y} and a vector of predicted categories \code{hx} the cumulative LGD accuarcy ratio (CLAR) according to Ozdemir and Miu 2009.
}
\examples{
clar(rep(1:5,each=3),c(3,3,3,rep(2:5,each=3)))
}
\references{
Ozdemir, B., Miu, P., 2009. Basel II Implementation. A Guide to Developing and Validating a Compliant Internal Risk Rating System. McGraw-Hill, USA.
}
